requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
plaid-python>=9.0.0
tavily-python>=0.3.0
groq>=0.4.0
langgraph>=0.0.55
langchain>=0.1.0
langchain-groq>=0.1.0
pydantic>=2.0.0
pandas>=2.0.0
requests>=2.31.0
sentence-transformers>=2.2.0
plotly>=5.17.0
